pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.0",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-supervisor>=0.0.29",
    "langgraph-swarm>=0.0.14",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
